Move all models (created by Cesar) to /src/app/common/models.
