+++ src/app/cost-centers/cost-centers.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, ViewChild, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
 
 import { CostCentersService } from './cost-centers.service';
-import { CostCenter } from './cost-center.model';
+import { CostCenter } from '../common/models/cost-center.model';
 import { EditCostCenterComponent } from './edit-cost-center/edit-cost-center.component';
 
 @Component({
