+++ src/app/cost-centers/cost-centers.service.ts
@@ -5,7 +5,7 @@ import { map } from 'rxjs/operators';
 
 import { RequestService } from '../_services/request.service';
 import { DataHelperService } from '../_services/data.helper.service';
-import { CostCenter } from './cost-center.model';
+import { CostCenter } from '../common/models/cost-center.model';
 
 @Injectable({
   providedIn: 'root'
