--- src/app/users/cost-center.model.ts
@@ -1,9 +1,0 @@
-export class CostCenter {
-    public id: number;
-    public name: string;
-
-    constructor(id: number, name: string) {
-        this.id = id;
-        this.name = name;
-    }
-}
