+++ src/app/users/user-edit/user-edit-resolver.service.ts
@@ -1,12 +1,15 @@
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { User } from '../user.model';
-import { UsersService } from '../users.service';
-import { Department } from '../department.model';
-import { Role } from '../role.model';
-import { forkJoin, Observable, Subject } from 'rxjs';
+import { forkJoin, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
+
+import { UsersService } from '../users.service';
+import { DepartmentsService } from 'src/app/departments/departments.service';
+import { CostCentersService } from 'src/app/cost-centers/cost-centers.service';
 import { DataHelperService } from 'src/app/_services/data.helper.service';
+import { User } from '../../common/models/user.model';
+import { Department } from '../../common/models/department.model';
+import { Role } from '../../common/models/role.model';
 
 interface UserEditData {
   departments: Department[],
@@ -16,12 +19,17 @@ interface UserEditData {
 
 @Injectable({ providedIn: 'root' })
 export class UserEditResolverService implements Resolve<UserEditData> {
-  constructor(private userService: UsersService, private dataHelper: DataHelperService) {}
+  constructor(
+    private userService: UsersService,
+    private costCentersService: CostCentersService,
+    private departmentsService: DepartmentsService,
+    private dataHelper: DataHelperService
+  ) {}
 
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
     let queries : Observable<any>[] = [
-      this.userService.fetchDepartments(),
-      this.userService.fetchCostCenters(),
+      this.departmentsService.fetchAll(),
+      this.costCentersService.fetchAll(),
       this.userService.fetchRoles(),
     ];
